+++ include/kvm/virtio.h
@@ -225,14 +225,14 @@ struct virtio_ops {
 	int (*signal_config)(struct kvm *kvm, struct virtio_device *vdev);
 	void (*notify_status)(struct kvm *kvm, void *dev, u32 status);
 	int (*init)(struct kvm *kvm, void *dev, struct virtio_device *vdev,
-		    int device_id, int subsys_id, int class, u32 addr, u8 irq);
+		    int device_id, int subsys_id, int class, u32 addr, u32 irq);
 	int (*exit)(struct kvm *kvm, struct virtio_device *vdev);
 	int (*reset)(struct kvm *kvm, struct virtio_device *vdev);
 };
 
 int __must_check virtio_init(struct kvm *kvm, void *dev, struct virtio_device *vdev,
 			     struct virtio_ops *ops, enum virtio_trans trans,
-			     int device_id, int subsys_id, int class, u32 addr, u8 irq);
+			     int device_id, int subsys_id, int class, u32 addr, u32 irq);
 const char* virtio_trans_name(enum virtio_trans trans);
 void virtio_init_device_vq(struct kvm *kvm, struct virtio_device *vdev,
 			   struct virt_queue *vq, size_t nr_descs);
